requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Natural language processing
nltk>=3.8
spacy>=3.6.0
textstat>=0.7.3
newspaper3k>=0.2.8
feedparser>=6.0.10

# Web interface
streamlit>=1.28.0

# Web scraping and HTTP requests
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Reddit API integration
praw>=7.7.0

# Wikipedia API integration
wikipedia-api>=0.6.0

# Configuration and environment management
python-dotenv>=1.0.0

# Utilities and visualization
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and deployment
psutil>=5.9.0

# Testing
pytest>=7.4.0

# Enhanced feature extraction
vaderSentiment>=3.3.2
wordcloud>=1.9.2

# External dataset integration (optional)
# kaggle>=1.5.16  # Uncomment if you want to use Kaggle datasets
# datasets>=2.14.0  # Uncomment if you want to use HuggingFace datasets

# Advanced ML models
xgboost>=1.7.0
lightgbm>=4.0.0

# Optional: Production monitoring and error tracking
# Uncomment these for production deployments
# sentry-sdk>=1.32.0
# prometheus-client>=0.17.0
# redis>=4.6.0